Move external view types to common
